+++ CHANGES.md
@@ -15,6 +15,7 @@ Change Log
 * Leaflet attribution can be set through `config.leafletAttributionPrefix`. Attribution HTML string to show on Leaflet maps. Will use Leaflet's default if undefined. To hide Leaflet attribution - set `leafletAttributionPrefix:""`
 * Re-add missing `helpPanel.mapUserGuide` translation string
 * Fix `sortMembersBy` for child `Groups` and `References`
+* Add progress of 3DTiles data source loading to Progress Bar
 * [The next improvement]
 
 #### 8.2.3 - 2022-04-22